+++ CHANGELOG.md
@@ -4,6 +4,11 @@
   Placeholder for the next version (at the beginning of the line):
   ## **WORK IN PROGRESS**
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Fix issue with conflicting daikin and Sonos device names
+
 ## 0.1.42-alpha.6 (2022-07-12)
 
 * (theimo1221) Prevent infinity loop on daikin reconnect